Implemented SaveRam storage on separate .sram file with the same name as the .nes or .zip rom file.

+++ NesCore/Storage/CartridgeMapAxRom.cs
@@ -36,7 +36,7 @@ namespace NesCore.Storage
                     return Cartridge.ProgramRom[programBank * 0x8000 + address - 0x8000];
 
                 if (address >= 0x6000)
-                    return Cartridge.SaveRam[address - 0x6000];
+                    return Cartridge.SaveRam[(ushort)(address - 0x6000)];
 
                 throw new Exception("Unhandled " + Name + " mapper read at address: " + Hex.Format(address));
             }
@@ -53,7 +53,7 @@ namespace NesCore.Storage
                         ? MirrorMode.Single1 : MirrorMode.Single0;
                 }
                 else if (address >= 0x6000)
-                    Cartridge.SaveRam[address - 0x6000] = value;
+                    Cartridge.SaveRam[(ushort)(address - 0x6000)] = value;
                 else
                     throw new Exception("Unhandled " + Name + " mapper write at address: " + Hex.Format(address));
             }
