+++ NesCore/Storage/CartridgeMapMmc1.cs
@@ -47,7 +47,7 @@ namespace NesCore.Storage
                     return Cartridge.ProgramRom[programBankOffsets[bank] + offset];
                 }
                 else if (address >= 0x6000)
-                    return Cartridge.SaveRam[address - 0x6000];
+                    return Cartridge.SaveRam[(ushort)(address - 0x6000)];
                 else
                     throw new Exception("Unhandled " + Name + " mapper read at address: " + Hex.Format(address));
             }
@@ -66,7 +66,7 @@ namespace NesCore.Storage
                 }
                 else if (address >= 0x6000)
                 {
-                    Cartridge.SaveRam[address - 0x6000] = value;
+                    Cartridge.SaveRam[(ushort)(address - 0x6000)] = value;
                 }
             }
         }
