+++ NesCore/Storage/CartridgeMapMmc3.cs
@@ -52,7 +52,7 @@ namespace NesCore.Storage
                     return Cartridge.ProgramRom[programBankOffsets[bank] + offset];
                 }
                 else if (address >= 0x6000)
-                    return Cartridge.SaveRam[address - 0x6000];
+                    return Cartridge.SaveRam[(ushort)(address - 0x6000)];
                 else
                     throw new Exception("Unhandled " + Name + " mapper write at address: " + Hex.Format(address));
             }
@@ -68,7 +68,7 @@ namespace NesCore.Storage
                 else if (address >= 0x8000)
                     WriteRegister(address, value);
                 else if (address >= 0x6000)
-                    Cartridge.SaveRam[address - 0x6000] = value;
+                    Cartridge.SaveRam[(ushort)(address - 0x6000)] = value;
                 else
                     throw new Exception("Unhandled " + Name + " mapper write at address: " + Hex.Format(address));
             }
