+++ NesCore/Storage/CartridgeMapNRom.cs
@@ -47,7 +47,7 @@ namespace NesCore.Storage
                 }
 
                 if (address >= 0x6000)
-                    return Cartridge.SaveRam[address - 0x6000];
+                    return Cartridge.SaveRam[(ushort)(address - 0x6000)];
 
                 throw new Exception("Unhandled " + Name + " mapper read at address: " + Hex.Format(address));
             }
@@ -64,7 +64,7 @@ namespace NesCore.Storage
                 }
                 else if (address >= 0x6000)
                 { 
-                    Cartridge.SaveRam[address - 0x6000] = value;
+                    Cartridge.SaveRam[(ushort)(address - 0x6000)] = value;
                 }
                 else
                     throw new Exception("Unhandled " + Name + " mapper write at address: " + Hex.Format(address));
